fix(git): Ensure origin/HEAD is set
When we query for the default branch, we assume `origin/HEAD` exists. That said this ref is only set and needed when you clone a repo and GitHub's `action/checkout` seems to be doing something weird that doesn't set this, causing issues like https://github.com/getsentry/sentry-javascript/runs/2730108174?check_suite_focus=true#step:3:10. This patch should fix the issue.